chore: move common utility functions to `frappe.utils`
