+++ frappe/utils/telemetry/pulse/app_heartbeat_event.py
@@ -1,7 +1,7 @@
 import frappe
+from frappe.utils import get_app_version, get_frappe_version
 
 from .client import capture, is_enabled
-from .utils import get_app_version, get_frappe_version
 
 
 def capture_app_heartbeat(app):
